allow multiple grants for one organisation in one project

+++ app/models/grant.rb
@@ -8,7 +8,7 @@ class Grant
   field :offer_cost, Vocabulary::TSBDEF.offerCost, datatype: RDF::XSD.integer
   field :offer_grant, Vocabulary::TSBDEF.offerGrant, datatype: RDF::XSD.integer
   field :offer_percentage, Vocabulary::TSBDEF.offerPercentage, datatype: RDF::XSD.decimal
-  field :payments_to_date, Vocabulary::TSBDEF.paymentsToDate, datatype: RDF::XSD.integer
+  field :payments_to_date, Vocabulary::TSBDEF.paymentsToDate, datatype: RDF::XSD.decimal
 
   # links
   linked_to :paid_to_organization, Vocabulary::TSBDEF.paidTo, class_name: 'Organization'
